==> notebooks/3.0 -zp-building-predictive-model.ipynb <==
## Building Predictive Models

import pandas as pd
import os
import numpy as np

### Import Data

# set the path of the processed data
processed_data_path = os.path.join(os.path.pardir, 'data', 'processed')
train_file_path = os.path.join(processed_data_path, 'train.csv')
test_file_path = os.path.join(processed_data_path, 'test.csv')

train_df = pd.read_csv(train_file_path, index_col = 'PassengerId')
test_df = pd.read_csv(test_file_path, index_col = 'PassengerId')

train_df.info()

test_df.info()

### Data Preparation

X = train_df.loc[:,'Age':].values.astype('float')
y = train_df['Survived'].ravel()

type(X)

type(y)

X.shape

y.shape

# train/test split
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 0)
print(X_train.shape, y_train.shape)
print(X_test.shape, y_test.shape)

# average survival in train and test
print('Mean Survival Rate in train: {0:.3f}'.format(np.mean(y_train)))
print('Mean Survival Rate in test: {0:.3f}'.format(np.mean(y_test)))

#### Check scikit-learn version

import sklearn

sklearn.__version__

### Baseline Model

# import function
from sklearn.dummy import DummyClassifier

# create model
model_dummy = DummyClassifier(strategy = 'most_frequent', random_state = 0)

# train model
model_dummy.fit(X_train, y_train)

print('Score for Baseline Model: {0:.2f}'.format(model_dummy.score(X_test, y_test))) # default score is Accuracy

# performance metrics
from sklearn.metrics import accuracy_score, confusion_matrix, precision_score, recall_score

# accuracy score
print('Accuracy for Baseline Mode: {0:.2f}'.format(accuracy_score(y_test, model_dummy.predict(X_test))))

# confusion matrix
print('Confusion Matrix for Baseline Mode:\n {0}'.format(confusion_matrix(y_test, model_dummy.predict(X_test))))

# precision and recall scores
print('Precision for Baseline Model: {0:.2f}'.format(precision_score(y_test, model_dummy.predict(X_test))))
print('Recall for Baseline Mode: {0:.2f}'.format(recall_score(y_test, model_dummy.predict(X_test))))

### First Kaggle Submission

# converting to the matrix
test_X = test_df.values.astype('float')

# get predictions
predictions = model_dummy.predict(test_X)

df_submission = pd.DataFrame({'PassengerId' : test_df.index, 'Survived' : predictions})

df_submission.head()

submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
submission_file_path = os.path.join(submission_data_path, '01_dummy.csv')

df_submission.to_csv(submission_file_path, index = False)

def get_submission_file(model, filename):
    # converting to the matrix
    test_X = test_df.values.astype('float')
    # get predictions
    predictions = model.predict(test_X)
    # submission DataFrame
    df_submission = pd.DataFrame({'PassengerId' : test_df.index, 'Survived' : predictions})
    # submission file
    submission_data_path = os.path.join(os.path.pardir, 'data', 'external')
    submission_file_path = os.path.join(submission_data_path, filename)
    # write to file
    df_submission.to_csv(submission_file_path, index = False)

# get submission file
get_submission_file(model_dummy, '01_dummy.csv')

### Logistic Regression

# import function
from sklearn.linear_model import LogisticRegression

# create model
model_lr_1 = LogisticRegression(random_state = 0)

# train model
model_lr_1.fit(X_train, y_train)

# evaluate model
print('Score for Logistic Regression - V1: {0:.2f}'.format(model_lr_1.score(X_test, y_test)))

# model coefficients
model_lr_1.coef_

### Second Kaggle Submission

get_submission_file(model_lr_1, '02_lr.csv')

# Part 2

## Hyperparameter Optimization

# base model
model_lr = LogisticRegression(random_state = 0, max_iter = 10000, solver = 'liblinear')

from sklearn.model_selection import GridSearchCV

parameters = {'C' : [1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty' : ['l1','l2']}
clf = GridSearchCV(model_lr, param_grid = parameters, cv = 3)

clf.fit(X_train, y_train)

clf.best_params_

print('Best Score: {0:.2f}'.format(clf.best_score_))

# evaluate model
print('Score for Logistic Regression - V2: {0:.2f}'.format(clf.score(X_test, y_test)))

## Third Kaggle Submission

# get submission file
get_submission_file(clf, '03_lr.csv')

### Feature Normalization and Standardization

from sklearn.preprocessing import MinMaxScaler, StandardScaler

#### Feature Normalization

# feature normalization
scaler = MinMaxScaler()
X_train_scaled = scaler.fit_transform(X_train)

X_train_scaled[:,0].min(), X_train_scaled[:,0].max()

# normalize test data
X_test_scaled = scaler.transform(X_test)

#### Feature Standardization

# feature standardization
scaler = StandardScaler()
X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

#### Create Model After Standardization

# base model
model_lr = LogisticRegression(random_state = 0, max_iter = 10000, solver = 'liblinear')
parameters = {'C' : [1.0, 10.0, 50.0, 100.0, 1000.0], 'penalty' : ['l1', 'l2']}
clf = GridSearchCV(model_lr, param_grid = parameters, cv = 3)
clf.fit(X_train_scaled, y_train)

clf.best_score_

# evaluate model
print('Score for Logistic Regression - V3: {0:.2f}'.format(clf.score(X_test_scaled, y_test)))

### Model Persistence

# import pickle library
import pickle

# create the file paths
model_file_path = os.path.join(os.path.pardir, 'models', 'lr_model.pkl')
scaler_file_path = os.path.join(os.path.pardir, 'models', 'lr_scaler.pkl')

# open the files to write
model_file_pickle = open(model_file_path, 'wb')
scaler_file_pickle = open(scaler_file_path, 'wb')

# persist the model and scaler
pickle.dump(clf, model_file_pickle)
pickle.dump(scaler, scaler_file_pickle)

# close the file
model_file_pickle.close()
scaler_file_pickle.close()

#### Load the Persisted File

# open files in read mode
model_file_pickle = open(model_file_path, 'rb')
scaler_file_pickle = open(scaler_file_path, 'rb')
# load files
clf_loaded = pickle.load(model_file_pickle)
scaler_loaded = pickle.load(scaler_file_pickle)
# close files
model_file_pickle.close()
scaler_file_pickle.close()

clf_loaded

scaler_loaded

# transform the test data using loaded scaler object
X_test_scaled = scaler_loaded.transform(X_test)
# calculate the score using the loaded model object
print('Score for Persisted Logistic Regression: {0:.2f}'.format(clf_loaded.score(X_test_scaled, y_test)))
